+++ CHANGELOG.md
@@ -14,6 +14,7 @@ If you're using an image that comes with an existing system Python installation,
 
 ### Added
 
+- Added support for choosing clusters based on tags (e.g. `gantry run --tag='storage:weka' ...`).
 - Added `--python-manager` option to `gantry run` command.
 - Added `--system-python` flag to `gantry run` command.
 - Added `--python-venv` flag to `gantry run` command.
@@ -25,6 +26,7 @@ If you're using an image that comes with an existing system Python installation,
 
 ### Changed
 
+- Gantry will automatically apply a `--tag='storage:weka'` cluster filter rule when you use a `--weka` option.
 - Renamed `--conda` option to `--conda-file` in `gantry run` command.
 - Renamed `--venv` option to `--conda-env` in `gantry run` command.
 - Renamed `--python-version` to `--default-python-version` in `gantry run` command.